# platform-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hyperfleet Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /authz/check:
    post:
      summary: Check authorization
      description: |
        Evaluates an authorization request against the account's policies
        and returns an ALLOW or DENY decision. Requires a provisioned account.
      operationId: checkAuthorization
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckAuthorizationRequest"
      responses:
        "200":
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckAuthorizationResponse"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - account not provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Authorization - Policy Management
  /authz/policies:
    post:
      summary: Create a policy
      description: |
        Creates a new Cedar policy. The policy body is raw Cedar policy text,
        not a structured object.
        Policies must be attached to users or groups to take effect.
      operationId: createPolicy
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Bad request - invalid policy format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - account not provisioned or caller not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List policies
      description: Returns a list of all policies for the account.
      operationId: listPolicies
      tags:
        - Authorization
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /authz/policies/{id}:
    get:
      summary: Get a policy
      description: Returns a specific policy by ID.
      operationId: getPolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Policy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a policy
      description: |
        Updates an existing policy. All attachments using this policy will be
        updated with the new policy content.
      operationId: updatePolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
      responses:
        "200":
          description: Policy updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Bad request - invalid policy format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a policy
      description: |
        Deletes a policy. Returns 409 if the policy has existing attachments.
      operationId: deletePolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Policy deleted successfully
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Policy has existing attachments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Authorization - Group Management
  /authz/groups:
    post:
      summary: Create a group
      description: Creates a new group that can contain multiple users.
      operationId: createGroup
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List groups
      description: Returns a list of all groups for the account.
      operationId: listGroups
      tags:
        - Authorization
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /authz/groups/{id}:
    get:
      summary: Get a group
      description: Returns a specific group by ID.
      operationId: getGroup
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a group
      description: |
        Deletes a group and all its member associations.
        Attachments to this group will also be deleted.
      operationId: deleteGroup
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Group deleted successfully
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /authz/attachments/{id}:
    delete:
      summary: Delete an attachment
      description: |
        Detaches a policy from a user or group. The Cedar policy in
        Amazon Verified Permissions will be deleted.
      operationId: deleteAttachment
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Attachment ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Attachment deleted successfully
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Authorization - Admin Management
...
  /authz/admins/{arn}:
    delete:
      summary: Remove an admin
      description: Removes a principal from the admin list.
      operationId: removeAdmin
      tags:
        - Authorization
      parameters:
        - name: arn
          in: path
          required: true
          description: Admin's ARN (URL-encoded)
          schema:
            type: string
      responses:
        "204":
          description: Admin removed successfully
        "403":
          description: Forbidden - caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ZOA Trusted Actions
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - kind
        - code
        - reason
      properties:
        kind:
          type: string
          description: Resource kind
          example: Error
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
...
    CheckAuthorizationRequest:
      type: object
      description: Request body for checking authorization
      required:
        - principal
        - action
        - resource
      properties:
        principal:
          type: string
          description: Principal ARN making the request
        action:
          type: string
          description: "Action being performed (e.g., rosa:CreateCluster)"
        resource:
          type: string
          description: "Resource ARN (e.g., arn:aws:rosa:us-west-2:123456789012:cluster/*)"
        context:
          type: object
          description: Additional context for the authorization decision
          additionalProperties: true
        resourceTags:
          type: object
          description: Tags on the resource
          additionalProperties:
            type: string
    CheckAuthorizationResponse:
      type: object
      description: Authorization decision response
      required:
        - kind
        - decision
      properties:
        kind:
          type: string
          example: AuthorizationDecision
        decision:
          type: string
          description: Authorization decision
          enum: [ALLOW, DENY]
        reason:
          type: string
          description: Reason for the decision
...